Move parser class to parser package
