+++ src/main/java/io/github/lavenderses/aws_app_config_openfeature_provider/AwsAppConfigClientService.java
@@ -7,8 +7,8 @@ import dev.openfeature.sdk.ErrorCode;
 import dev.openfeature.sdk.Reason;
 import dev.openfeature.sdk.Value;
 import io.github.lavenderses.aws_app_config_openfeature_provider.app_config_model.AppConfigValue;
-import io.github.lavenderses.aws_app_config_openfeature_provider.app_config_model.AppConfigValueParseException;
-import io.github.lavenderses.aws_app_config_openfeature_provider.app_config_model.AwsAppConfigParser;
+import io.github.lavenderses.aws_app_config_openfeature_provider.parser.AppConfigValueParseException;
+import io.github.lavenderses.aws_app_config_openfeature_provider.parser.AwsAppConfigParser;
 import io.github.lavenderses.aws_app_config_openfeature_provider.converter.AppConfigValueConverter;
 import io.github.lavenderses.aws_app_config_openfeature_provider.evaluation_value.ErrorEvaluationValue;
 import io.github.lavenderses.aws_app_config_openfeature_provider.evaluation_value.EvaluationValue;
