+++ src/main/java/io/github/lavenderses/aws_app_config_openfeature_provider/app_config_model/AppConfigBooleanValue.java
@@ -18,7 +18,7 @@ import static java.util.Objects.requireNonNull;
 @EqualsAndHashCode(callSuper = true)
 public final class AppConfigBooleanValue extends AppConfigValue<Boolean> {
 
-    AppConfigBooleanValue(
+    public AppConfigBooleanValue(
         @NotNull Boolean enabled,
         @NotNull Boolean value,
         @Language("json") @NotNull String jsonFormat
